requirements.txt
amqp==2.6.0
ansible==2.10.0
ansible-runner==1.4.6
ansible-vault==1.2.0
asgiref==3.3.0
celery==4.4.6
Django==3.1.3
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-datatables==0.5.2
hvac==0.10.5
itsdangerous==1.1.0
maya==0.6.1
mysql-connector-python==8.0.22
mysqlclient==2.0.1
pandas==1.1.4
pylint==2.7.2
pytz==2020.4
pytest==6.1.2
Pygments==2.7.2
requests==2.23.0
ruamel.yaml==0.16.10
sqlparse==0.4.1
urllib3==1.25.9
webservices==0.7
XlsxWriter==1.3.7